Move engine_scripts up, allow linting and formatting
