--- testing/visual-regression/backstop_data/engine_scripts/puppet/onReady.js
@@ -1,7 +1,0 @@
-module.exports = async (page, scenario, vp) => {
-    console.log('SCENARIO > ' + scenario.label);
-    await require('./overrideCSS')(page, scenario);
-    await require('./clickAndHoverHelper')(page, scenario);
-
-    // add more ready handlers here...
-};
