+++ testing/visual-regression/engine_scripts/cookies.json
