+++ testing/visual-regression/engine_scripts/imageStub.jpg
