+++ testing/visual-regression/engine_scripts/puppet/clickAndHoverHelper.js
@@ -1,14 +1,18 @@
 module.exports = async (page, scenario) => {
   const hoverSelector = scenario.hoverSelectors || scenario.hoverSelector;
   const clickSelector = scenario.clickSelectors || scenario.clickSelector;
-  const keyPressSelector = scenario.keyPressSelectors || scenario.keyPressSelector;
+  const keyPressSelector =
+    scenario.keyPressSelectors || scenario.keyPressSelector;
   const scrollToSelector = scenario.scrollToSelector;
   const postInteractionWait = scenario.postInteractionWait; // selector [str] | ms [int]
 
   if (keyPressSelector) {
     for (const keyPressSelectorItem of [].concat(keyPressSelector)) {
       await page.waitFor(keyPressSelectorItem.selector);
-      await page.type(keyPressSelectorItem.selector, keyPressSelectorItem.keyPress);
+      await page.type(
+        keyPressSelectorItem.selector,
+        keyPressSelectorItem.keyPress
+      );
     }
   }
 
@@ -32,7 +36,7 @@ module.exports = async (page, scenario) => {
 
   if (scrollToSelector) {
     await page.waitFor(scrollToSelector);
-    await page.evaluate(scrollToSelector => {
+    await page.evaluate((scrollToSelector) => {
       document.querySelector(scrollToSelector).scrollIntoView();
     }, scrollToSelector);
   }
