+++ testing/visual-regression/engine_scripts/puppet/onBefore.js
